+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.11.1] - 2020-06-30
 ### Added
 - `#[pyclass(unsendable)]`. [#1009](https://github.com/PyO3/pyo3/pull/1009)
 
@@ -433,7 +435,8 @@ Yanked
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/pyo3/pyo3/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/pyo3/pyo3/compare/v0.11.1...HEAD
+[0.11.0] https://github.com/pyo3/pyo3/compare/v0.11.0...v0.11.1
 [0.11.0] https://github.com/pyo3/pyo3/compare/v0.10.1...v0.11.0
 [0.10.1]: https://github.com/pyo3/pyo3/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/pyo3/pyo3/compare/v0.9.2...v0.10.0